+++ src/cms/preview-templates/ProjectPreview.tsx
@@ -2,11 +2,8 @@ import React, { FunctionComponent } from 'react'
 import { PreviewProps, getJSON } from './helpers'
 import { createPreview } from './Preview'
 import { ProjectPageComponent } from '../../components/PastProjects/Page'
-import {
-  ProjectData,
-  translateAndConvert,
-  ProjectDataOnDisk,
-} from '../../data/project'
+import { ProjectData, ProjectDataOnDisk } from '../../data/project'
+import { translateAndConvert } from '../../data/project/helpers'
 
 const Preview = createPreview<ProjectDataOnDisk, ProjectData>()
