+++ src/data/technology/index.tsx
@@ -2,7 +2,3 @@ export interface Technology {
   name: string
   link: string
 }
-
-export type StringifiedTechnologies = string[]
-
-type TechnologiesOnDisk = Technology[]
